Final Commit
Working
~low end ChessClub website~
___________
[Main]
Navigation bar with hidden values depending if the user is logged in or not.

Working Payments, Once purchased is press will be re-directed to payment page where user can input the amount they want to join a competition.

Working Admin panel that allows an admin to create, edit or delete users.

[Database stuff]

Working login & user registration linked up with database

Admin Panel can change values or add values in the user database.

[Tests]
Only able to get Web testing working due to the codecept version not being compatible with version of php.

None Working
__________________

[Main]

[ChessOnline webpage]

Wasn't able to get a working Chess game/puzzle working in the provided time scale.

[Profile webpage]

Is meant to show the competition a user has entered, but wasn't able to implement it properly.

Is meant to show Game History but as I was unable to get a working game it doesn't actually exist.

[Testing]
A lot of the testing options were not working due to compatibility issues.